requirements.txt
numpy~=1.21.5
pillow~=9.0.1
matplotlib~=3.5.1
torchvision~=0.12.0
pandas~=1.4.2
tqdm~=4.64.0